requirements.txt
requests
pytest
pytest-html
python-dotenv
jsonschema
backoff
Faker
